Adding restrictions to users endpoints for managaers and users

+++ src/main/java/org/example/controller/UserController.java
@@ -1,11 +1,13 @@
 package org.example.controller;
 
+import jakarta.servlet.http.HttpServletRequest;
 import org.example.Service.UserService;
 import org.example.model.User;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @RestController
@@ -18,12 +20,24 @@ public class UserController {
     }
 
     @GetMapping
-    public ResponseEntity<List<User>> getAllUsers(){
+    public ResponseEntity<List<User>> getAllUsers(HttpServletRequest request){
+        // Only manager can gat a list of all users
+        User sessionUser = (User) request.getAttribute("user");
+        if(sessionUser.getUserType().getId() != 1) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        }
+
         return ResponseEntity.ok(userService.findAllUsers());
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<User> getUserById(@PathVariable Long id) {
+    public ResponseEntity<User> getUserById(@PathVariable Long id, HttpServletRequest request) {
+        // Only manager and user itself can get user information
+        User sessionUser = (User) request.getAttribute("user");
+        if(sessionUser.getUserType().getId() != 1 && sessionUser.getId() != id) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        }
+
         return userService.findUserById(id)
                 .map(ResponseEntity::ok)
                 .orElse(ResponseEntity.notFound().build());
@@ -35,15 +49,31 @@ public class UserController {
         return ResponseEntity.status(HttpStatus.CREATED).body(savedUser);
     }
 
+
     @PutMapping("/{id}")
-    public ResponseEntity<User> updateUser(@PathVariable Long id, @RequestBody User userDetails){
+    public ResponseEntity<User> updateUser(@PathVariable Long id,
+                                           @RequestBody User userDetails,
+                                           HttpServletRequest request){
+        // Only manager and user itself can update user
+        User sessionUser = (User) request.getAttribute("user");
+        if(sessionUser.getUserType().getId() != 1 && sessionUser.getId() != id) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        }
+
         return userService.updateUser(id, userDetails)
                 .map(ResponseEntity::ok)
                 .orElse(ResponseEntity.notFound().build());
     }
 
+
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteUser(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteUser(@PathVariable Long id, HttpServletRequest request) {
+        // Only manager and user itself can delete user
+        User sessionUser = (User) request.getAttribute("user");
+        if(sessionUser.getUserType().getId() != 1 && sessionUser.getId() != id) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        }
+
         boolean deleted = userService.deleteUser(id);
         return deleted ? ResponseEntity.noContent().build() : ResponseEntity.notFound().build();
     }
